rename import/export methods
We make the following changes in the interests of conciseness,
consistency and clarity.

exportToJSONObject becomes exportJSON.
serializeBinary becomes exportBinary.
importMacaroon replaces importFromJSONObject and deserializeBinary but always imports a single macaroon only.
importMacaroons also replaces importFromJSONObject and deserializeBinary but alwaus imports an array of macaroons (it also allows import of a single macaroon, even though it always returns an array).